+++ cocos/rendering/post-process/passes/base-pass.ts
@@ -3,10 +3,25 @@ import { EDITOR } from 'internal:constants';
 import { Material } from '../../../asset/assets';
 import { Camera } from '../../../render-scene/scene';
 import { getCameraUniqueID } from '../../custom/define';
-import { Pipeline } from '../../custom/pipeline';
+import { BasicPipeline, Pipeline } from '../../custom/pipeline';
 import { passContext } from '../utils/pass-context';
+import { Format } from '../../../gfx';
+import { supportsRGBA16FloatTexture } from '../../define';
+import { macro } from '../../../core';
 
 let _BasePassID = 0;
+
+export function GetRTFormatBeforeToneMapping (ppl: BasicPipeline) {
+    const useFloatOutput = ppl.getMacroBool('CC_USE_FLOAT_OUTPUT');
+    return ppl.pipelineSceneData.isHDR && useFloatOutput && supportsRGBA16FloatTexture(ppl.device) ? Format.RGBA16F : Format.RGBA8;
+}
+export function ForceEnableFloatOutput (ppl: BasicPipeline) {
+    if (ppl.pipelineSceneData.isHDR && !ppl.getMacroBool('CC_USE_FLOAT_OUTPUT')) {
+        const supportFloatOutput = supportsRGBA16FloatTexture(ppl.device);
+        ppl.setMacroBool('CC_USE_FLOAT_OUTPUT', supportFloatOutput);
+        macro.ENABLE_FLOAT_OUTPUT = supportFloatOutput;
+    }
+}
 
 export abstract class BasePass {
     abstract name: string;
