+++ cocos/rendering/post-process/passes/forward-transparency-pass.ts
@@ -28,8 +28,7 @@ import { LightInfo, QueueHint, SceneFlags } from '../../custom';
 import { getCameraUniqueID } from '../../custom/define';
 import { Pipeline } from '../../custom/pipeline';
 import { passContext } from '../utils/pass-context';
-import { BasePass } from './base-pass';
-import { GetRTFormatBeforeToneMapping } from './setting-pass';
+import { BasePass, GetRTFormatBeforeToneMapping } from './base-pass';
 import { ShadowPass } from './shadow-pass';
 
 export class ForwardTransparencyPass extends BasePass {
